backend/requirements.txt
requests
beautifulsoup4
lxml
nltk
chromadb
sentence-transformers
fastapi[all]
google-generativeai
